privval: move proto aliases from api package
