+++ privval/signer_client.go
@@ -4,9 +4,9 @@ import (
 	"fmt"
 	"time"
 
-	pvproto "github.com/cometbft/cometbft/api/cometbft/privval"
 	cmtproto "github.com/cometbft/cometbft/api/cometbft/types"
 	"github.com/cometbft/cometbft/crypto"
+	pvproto "github.com/cometbft/cometbft/privval/proto"
 	cmterrors "github.com/cometbft/cometbft/types/errors"
 
 	cryptoenc "github.com/cometbft/cometbft/crypto/encoding"
