+++ privval/signer_requestHandler.go
@@ -4,10 +4,10 @@ import (
 	"fmt"
 
 	cryptoproto "github.com/cometbft/cometbft/api/cometbft/crypto/v1beta1"
-	pvproto "github.com/cometbft/cometbft/api/cometbft/privval"
 	cmtproto "github.com/cometbft/cometbft/api/cometbft/types"
 	"github.com/cometbft/cometbft/crypto"
 	cryptoenc "github.com/cometbft/cometbft/crypto/encoding"
+	pvproto "github.com/cometbft/cometbft/privval/proto"
 	"github.com/cometbft/cometbft/types"
 )
